Initial code for loading Terse Executables
This commit adds minimal support for loading Terse Executables. There
is still more work to be done such as handling relocations and other
information. I think we'll also need new EFI platforms for the PEI
stage so we can properly prototype the PEI module entrypoint and
apply structures to the TE image and section headers